==> notebooks/1_Data_Collection_API.ipynb <==
#Imports & Configuration

import requests

import pandas as pd

import numpy as np

import datetime

# See columns in the Dataframe

pd.set_option('display.max_columns', None)

pd.set_option('display.max_colwidth', None)

print("Libraries have been imported!")

# Specify the SpaceX API URL

spacex_url = "https://api.spacexdata.com/v4/launches/past"

print("Dataset URL successfully fetched")

# Translate return IDs to Understandable Column Names (RocketID to Booster Name, LaunchSite_Data to Launchpad Names, Payload_Data to Payload Mass and Orbit, Landing_Outcomes to True/False)

# Rocket ID to Booster Name

def getBoosterVersion(data):
    """Fetches the Booster name from the Rocket ID"""
    for x in data:
        if x['rocket']:
            response = requests.get("https://api.spacexdata.com/v4/rockets/"+str(x['rocket'])).json()
            x['rocket'] = response['name']

# LaunchSite_Data to Launchpad Name, Longitude and Latitude

def getLaunchSite(data):
    """Fetches the LAunchpad name, Longitude and Latitude"""
    for x in data:
        if x['launchpad']:
            response = requests.get("https://api.spacexdata.com/v4/launchpads/"+str(x['launchpad'])).json()
            x['launchpad'] = response['name']
            x['longitude'] = response['longitude']
            x['latitude'] = response['latitude']

# PayLoad_Data to Payload Mass and Orbit type

def getPayloadData(data):
    """Fetches the Payload Mass and Orbit type"""
    for x in data:
        if x['payloads']:
            response = requests.get("https://api.spacexdata.com/v4/payloads/"+str(x['payloads'][0])).json()
            x['payload_mass'] = response['mass_kg']
            x['orbit'] = response['orbit']

# Core_Data to Launch Outcome True/False

def getCoreData(data):
    """Fetches the Launch Outcome Data by True/False using If/Else"""
    for x in data:
        if x['cores']:
            core = x['cores'][0]
            if core['core'] != None:
                response = requests.get("https://api.spacexdata.com/v4/cores/"+str(core['core'])).json()
                x['block'] = response['block']
                x['reuse_count'] = response['reuse_count']
                x['serial'] = response['serial']
            else:
                x['block'] = None
                x['reuse_count'] = None
                x['serial'] = None
            x['outcome'] = core['landing_success']
            x['flights'] = core['flight']
            x['gridfins'] = core['gridfins']
            x['reused'] = core['reused']
            x['legs'] = core['legs']
            x['landingpad'] = core['landpad']

print("Column Names Defined")

# Fetch and Filter Data for Specific Launch Vehicle (Falcan 9) not other launches. Vehicle lists (https://www.spacex.com/launches) / https://www.spacex.com/vehicles/falcon-9

# Request the data
response = requests.get(spacex_url)
data = response.json()

# Convert received data to DataFrame
data = pd.DataFrame(data)

# Filter Launches for Falcon 9 (ID: 5e9d0d95eda69973a809d1ec)
data = data[data['rocket'] == '5e9d0d95eda69973a809d1ec']

# Filter for Launches that were successful to show
data = data[data['success'].notna()]

# Reset our list to JSON format for column names
data_json = data.to_dict(orient='records')

print(f"Filtered down to {len(data_json)} Falcon 9 launches.")

# Process the data from MANY API calls to look up Booster,LaunchSite,Payload and Core data

print("Starting Data Processing... Tasks 4/4 to complete... (This may take a while...)")

getBoosterVersion(data_json)
print("... Task 1: Booster data fetched")

getLaunchSite(data_json)
print("... Task 2: Launch Sites data fetched")

getPayloadData(data_json)
print("... Task 3: Payload data fetched")

getCoreData(data_json)
print("... Task 4: Core outcomes data fetched")

print("Tasks 4/4 Fully Processed!")

# Create dictionary / filter out data / save to file for later processing

launch_dict = {
    'FlightNumber': list(data['flight_number']),
    'Date': list(data['date_utc']),
    'BoosterVersion': [x['rocket'] for x in data_json],
    'PayloadMass': [x['payload_mass'] for x in data_json],
    'Orbit': [x['orbit'] for x in data_json],
    'LaunchSite': [x['launchpad'] for x in data_json],
    'Outcome': [x['outcome'] for x in data_json],
    'Flights': [x['flights'] for x in data_json],
    'GridFins': [x['gridfins'] for x in data_json],
    'Reused': [x['reused'] for x in data_json],
    'Legs': [x['legs'] for x in data_json],
    'LandingPad': [x['landingpad'] for x in data_json],
    'Block': [x['block'] for x in data_json],
    'ReusedCount': [x['reuse_count'] for x in data_json],
    'Serial': [x['serial'] for x in data_json],
    'Longitude': [x['longitude'] for x in data_json],
    'Latitude': [x['latitude'] for x in data_json]
}

# Convert to Dataframe
df = pd.DataFrame(launch_dict)

# Filter out any possible mention of Falcon 1
df = df[df['BoosterVersion'] != 'Falcon 1']

# Reset the Index
df.reset_index(drop=True, inplace=True)
df.head()

# Save our dictionary to CSV file for future processing
df.to_csv('1_data_collection.csv', index=False)
print("Data saved to 1_data_collection.csv")
