==> notebooks/2_Data_Wrangling.ipynb <==
# Load Data Libraries for Data Wrangling

import pandas as pd
import numpy as np

# Load dataset.csv created from notebook 1 - 1_data_collection.csv
df = pd.read_csv('1_data_collection.csv')

# Preview top 10 rows to check if dataset loaded correctly
df.head(10)

# Check for any missing values from dataset columns (Normal to have high number of missing data in Landingpad according to sources)
print(df.isnull().sum() / df.shape[0] * 100)

# Target Variable Cleaning of 'Outcome' column due to messy text rersults. Converting to Success (1) and Failure (0)
print(df['Outcome'].value_counts())

# ------- CODE DIDNT WORK CORRECTLY - KEEP AS EVIDENCE ---------

# Create Class Column using the converting criteria and filter out 'bad' outcomes (1 = Landed Successfully, 0 = Failed)
# bad_outcomes = {
#    'False ASDS', 
#    'False Ocean', 
#    'False RTLS', 
#    'None ASDS', 
#    'None None'
# }

# List for new filtered column
# landing_class = []

# for outcome in df['Outcome']:
#    if outcome in bad_outcomes:
#        landing_class.append(0) # Mark as Fail
#    else:
#        landing_class.append(1) # Mark as Success

# Add to the DataFrame
# df['Class'] = landing_class

# ---- SAMPLE OF FILTERED DATA FROM ABOVE CODE ----

#      Outcome  Class
# 0      NaN      1
# 1      NaN      1
# 2      NaN      1
# 3      NaN      1
# 4      NaN      1
# 5    False      1 <--- 
# 6      NaN      1
# 7      NaN      1
# 8     True      1 <---


# Fix - The Outcome column contains True, False, or NaN.
# 1 for True, and 0 for everything else (False or NaN).

# List for new filtered column and new logic
landing_class = []

for outcome in df['Outcome']:
    # Check if the value is explicitly True
    if outcome == True:
        landing_class.append(1) # Success
    else:
        landing_class.append(0) # Fail (False, NaN etc)

# Add to the dataframe
df['Class'] = landing_class

# Print our new outcome list
print(df[['Outcome', 'Class']].head(10))

# Calculate the success rate as a %age
success_rate = df['Class'].mean()
print(f"Success Rate: {success_rate:.2%}")

# Save the dataset into 2_data.wrangling.csv
df.to_csv('2_data_wrangling.csv', index=False)
print("Data wrangled and saved to 2_data_wrangling.csv")
