==> notebooks/3_EDA_SQL.ipynb <==
# Import libraries for SQL workings
import pandas as pd
import sqlite3

# Load our pre-made dataset (2_data_wrangling.csv)
df = pd.read_csv('2_data_wrangling.csv')
print("Dataset Loaded successfully!")

#Preview of top 5 rows to check loaded correctly
df.head(5)

# Create the SQL database
con = sqlite3.connect(":memory:")

# Convert the dataframe to a SQL table named: 'SPACEX'
df.to_sql("SPACEX", con, if_exists='replace', index=False, method="multi")
print("Database created and table 'SPACEX' now has data")

# Print top 5 rows of the new table to check row data and columns
verification_query = 'SELECT * FROM SPACEX LIMIT 5'

print(pd.read_sql(verification_query, con))

# Define the query function
def sql_query(query):
    return pd.read_sql(query, con)

# Ready for queries!
print("Database is now ready for queries!")

# Query 1 - Distinct Launchsites
query = """ 
SELECT DISTINCT LaunchSite FROM SPACEX
"""
sql_query(query)

# Query 2 -  Count flights and successes per booster version
query = """
SELECT BoosterVersion, 
       COUNT(*) as Total_Flights, 
       SUM(Class) as Successful_Landings
FROM SPACEX
GROUP BY BoosterVersion
ORDER BY Total_Flights DESC
LIMIT 10
"""
sql_query(query)

# Query 3 -  Calculate success rate per Orbit type
query = """
SELECT Orbit, 
       COUNT(*) as Flights, 
       SUM(Class) as Successes,
       (CAST(SUM(Class) AS FLOAT) / COUNT(*)) * 100 as Success_Rate
FROM SPACEX
GROUP BY Orbit
ORDER BY Success_Rate DESC
"""
sql_query(query)

# Query 4 -  Average Payload Mass for Success (1) vs Fail (0)
query = """
SELECT Class, AVG(PayloadMass) as Avg_Payload_kg
FROM SPACEX
GROUP BY Class
"""
sql_query(query)

# Close the connection as standard practice!
con.close()
print("Database connection closed successfully.")
